+++ test/feature/dummy-feature.js
@@ -6,7 +6,9 @@ const groupsSource = factory.resource('groups.bpmn');
 
 Feature('Dummy', () => {
   Scenario('Group of elements with categories', () => {
-    let context, definition;
+    let context;
+    /** @type {import('bpmn-elements').Definition} */
+    let definition;
 
     let ended;
     When('a source with groups is ran', async () => {
@@ -14,6 +16,10 @@ Feature('Dummy', () => {
       definition = new Definition(context);
       ended = definition.waitFor('end');
       definition.run();
+    });
+
+    And('category can be retrieved', () => {
+      expect(definition.getElementById('Category_1275ejz')?.placeholder).to.be.true;
     });
 
     Then('it runs to end', () => {